QuickFit3/table:   - added special column with increasing number (1,2,3,...)   - axis labels can be set from table columns   - bar-charts now support errors   - date/time/date-time axis labels work properly now   - added parser functions colsavg(), colsstd(), colsmedian(), ..., which calculate a column of averages for each row in a set of columns
QuickFit3/fitfunctions_general: 
  - added centered polynomial
  - added Gaussian distribution function

QuickFit3/lib:
  - QFTableModel: now supports up to 2^32 rows and columns

QuickFit3/lib/math_parser:
  - added functions cumsum(), cumprod(), diff(), mad(), nmad(), indexedmad(), indexednmad(), trapez()/trapz()

git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3215 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/interfaces/qfrdrtableinterface.h
@@ -13,23 +13,23 @@ class QFRDRTableInterface {
 
 
         /** \brief retuns the data stored in the table at a given position */
-        virtual QVariant tableGetData(quint16 row, quint16 column) const=0;
+        virtual QVariant tableGetData(quint32 row, quint32 column) const=0;
         /** \brief sets the data in the table at a given position, if the cell does not exist, the table is expanded! */
-        virtual void tableSetData(quint16 row, quint16 column, const QVariant& data)=0;
+        virtual void tableSetData(quint32 row, quint32 column, const QVariant& data)=0;
         /** \brief sets a whole column in a table  */
-        virtual void tableSetColumnData(quint16 column, const QList<QVariant>& data)=0;
+        virtual void tableSetColumnData(quint32 column, const QList<QVariant>& data)=0;
         /** \brief returns a whole column in a table  */
-        virtual QList<QVariant>  tableGetColumnData(quint16 column)=0;
+        virtual QList<QVariant>  tableGetColumnData(quint32 column)=0;
         /** \brief sets the column title */
-        virtual void tableSetColumnTitle( quint16 column, const QString& data)=0;
+        virtual void tableSetColumnTitle( quint32 column, const QString& data)=0;
         /** \brief returns a column's title */
-        virtual QString tableGetColumnTitle(quint16 column) const=0;
+        virtual QString tableGetColumnTitle(quint32 column) const=0;
         /** \brief retuns \c true if the table implementation supports math expressions in the cells */
         virtual bool tableSupportsExpressions() const=0;
         /** \brief sets the math expression in a specific cell */
-        virtual void tableSetExpression(quint16 row, quint16 column, const QString& expression=QString(""))=0;
+        virtual void tableSetExpression(quint32 row, quint32 column, const QString& expression=QString(""))=0;
         /** \brief returns the math expression in a specific cell */
-        virtual QString tableGetExpression(quint16 row, quint16 column) const=0;
+        virtual QString tableGetExpression(quint32 row, quint32 column) const=0;
         /** \brief reevaluates all math expressions */
         virtual void tableReevaluateExpressions()=0;
         /** \brief returns the number of columns in the table */
@@ -39,15 +39,15 @@ class QFRDRTableInterface {
         /** \brief returns \c true if the table is read-only */
         virtual bool tableIsReadonly() const=0;
         /** \brief sets a whole column in a table  */
-        virtual void tableSetColumnDataAsDouble(quint16 column, const QVector<double>& data)=0;
+        virtual void tableSetColumnDataAsDouble(quint32 column, const QVector<double>& data)=0;
         /** \brief returns a whole column in a table converted to  numbers  */
-        virtual QVector<double> tableGetColumnDataAsDouble(quint16 column)=0;
+        virtual QVector<double> tableGetColumnDataAsDouble(int column)=0;
         /** \brief set the expression used to calculate the contents of the given column  */
-        virtual void tableSetColumnExpression(quint16 column, const QString& expression=QString(""))=0;
+        virtual void tableSetColumnExpression(quint32 column, const QString& expression=QString(""))=0;
         /** \brief set the comment of the given column  */
-        virtual void tableSetColumnComment(quint16 column, const QString& comment=QString(""))=0;
+        virtual void tableSetColumnComment(quint32 column, const QString& comment=QString(""))=0;
         /** \brief set the comment for the given cell (row, column)  */
-        virtual void tableSetComment(quint16 row, quint16 column, const QString& comment=QString(""))=0;
+        virtual void tableSetComment(quint32 row, quint32 column, const QString& comment=QString(""))=0;
 
 };
